refactor driver/datagram
