+++ autd3-firmware-emulator/tests/op/info.rs
@@ -3,7 +3,7 @@ use autd3_driver::{
     error::AUTDInternalError,
     firmware::{
         cpu::TxDatagram,
-        operation::{OperationGenerator, OperationHandler},
+        operation::{FirmwareVersionType, OperationGenerator, OperationHandler},
         version::FirmwareVersion,
     },
 };
@@ -13,6 +13,8 @@ use crate::{create_geometry, send};
 
 #[test]
 fn send_firminfo() -> anyhow::Result<()> {
+    use FirmwareVersionType::*;
+
     const EMULATOR_BIT: u8 = 1 << 7;
 
     let geometry = create_geometry(1);
@@ -27,26 +29,26 @@ fn send_firminfo() -> anyhow::Result<()> {
         assert!(cpu.reads_fpga_state());
     }
 
-    send(&mut cpu, FetchFirmInfo::CPUMajor, &geometry, &mut tx)?;
+    send(&mut cpu, CPUMajor, &geometry, &mut tx)?;
     assert_eq!(FirmwareVersion::LATEST_VERSION_NUM_MAJOR, cpu.rx().data());
     assert!(!cpu.reads_fpga_state());
 
-    send(&mut cpu, FetchFirmInfo::CPUMinor, &geometry, &mut tx)?;
+    send(&mut cpu, CPUMinor, &geometry, &mut tx)?;
     assert_eq!(FirmwareVersion::LATEST_VERSION_NUM_MINOR, cpu.rx().data());
     assert!(!cpu.reads_fpga_state());
 
-    send(&mut cpu, FetchFirmInfo::FPGAMajor, &geometry, &mut tx)?;
+    send(&mut cpu, FPGAMajor, &geometry, &mut tx)?;
     assert!(!cpu.reads_fpga_state());
 
-    send(&mut cpu, FetchFirmInfo::FPGAMinor, &geometry, &mut tx)?;
+    send(&mut cpu, FPGAMinor, &geometry, &mut tx)?;
     assert_eq!(FirmwareVersion::LATEST_VERSION_NUM_MINOR, cpu.rx().data());
     assert!(!cpu.reads_fpga_state());
 
-    send(&mut cpu, FetchFirmInfo::FPGAFunctions, &geometry, &mut tx)?;
+    send(&mut cpu, FPGAFunctions, &geometry, &mut tx)?;
     assert_eq!(EMULATOR_BIT, cpu.rx().data());
     assert!(!cpu.reads_fpga_state());
 
-    send(&mut cpu, FetchFirmInfo::Clear, &geometry, &mut tx)?;
+    send(&mut cpu, Clear, &geometry, &mut tx)?;
     assert!(cpu.reads_fpga_state());
 
     Ok(())
@@ -58,7 +60,7 @@ fn invalid_info_type() -> anyhow::Result<()> {
     let mut cpu = CPUEmulator::new(0, geometry.num_transducers());
     let mut tx = TxDatagram::new(geometry.num_devices());
 
-    let d = FetchFirmInfo::CPUMajor;
+    let d = FirmwareVersionType::CPUMajor;
     let (op, op_null) = d.operation_generator(&geometry)?.generate(&geometry[0]);
 
     OperationHandler::pack(&mut [(op, op_null)], &geometry, &mut tx, usize::MAX)?;
